+++ src/main/java/com/bootcamp/schedulemanagementjpaapp/dto/response/WeatherResponseDto.java
@@ -1,4 +1,4 @@
-package com.bootcamp.schedulemanagementjpaapp.dto;
+package com.bootcamp.schedulemanagementjpaapp.dto.response;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
